huge refactor nd updates

+++ packages/contracts/script/RegisterAgent.s.sol
@@ -9,7 +9,7 @@ import {IIdentityRegistry} from "../src/registry/interfaces/IIdentityRegistry.so
  * @title RegisterAgent
  * @notice Helper script to register agents with metadata
  * @dev Makes it easy to register agents from command line
- * 
+ *
  * Usage:
  *   forge script script/RegisterAgent.s.sol:RegisterAgent \
  *     --rpc-url http://localhost:8545 \
@@ -22,14 +22,13 @@ contract RegisterAgent is Script {
         string memory agentUri = vm.envOr("AGENT_URI", string("ipfs://QmDefaultAgent"));
         string memory agentName = vm.envOr("AGENT_NAME", string("My AI Agent"));
         string memory agentType = vm.envOr("AGENT_TYPE", string("general"));
-        
+
         // Get deployer key
         uint256 deployerPrivateKey = vm.envOr(
-            "DEPLOYER_PRIVATE_KEY",
-            uint256(0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80)
+            "DEPLOYER_PRIVATE_KEY", uint256(0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80)
         );
         address deployer = vm.addr(deployerPrivateKey);
-        
+
         console.log("==========================================");
         console.log("Registering AI Agent");
         console.log("==========================================");
@@ -39,12 +38,12 @@ contract RegisterAgent is Script {
         console.log("Agent Name:", agentName);
         console.log("Agent Type:", agentType);
         console.log("");
-        
+
         // Load registry from deployment if not provided
         if (registryAddress == address(0)) {
             string memory network = vm.envOr("NETWORK", string("localnet"));
             string memory path = string.concat("deployments/", network, "/liquidity-system.json");
-            
+
             try vm.readFile(path) returns (string memory json) {
                 registryAddress = payable(vm.parseJsonAddress(json, ".identityRegistry"));
                 console.log("Loaded registry from deployment:", registryAddress);
@@ -52,30 +51,21 @@ contract RegisterAgent is Script {
                 revert("Registry address not found. Set IDENTITY_REGISTRY env var.");
             }
         }
-        
+
         IdentityRegistry registry = IdentityRegistry(registryAddress);
-        
+
         vm.startBroadcast(deployerPrivateKey);
-        
+
         // Register agent with metadata
         IIdentityRegistry.MetadataEntry[] memory metadata = new IIdentityRegistry.MetadataEntry[](3);
-        metadata[0] = IIdentityRegistry.MetadataEntry({
-            key: "name",
-            value: abi.encode(agentName)
-        });
-        metadata[1] = IIdentityRegistry.MetadataEntry({
-            key: "type",
-            value: abi.encode(agentType)
-        });
-        metadata[2] = IIdentityRegistry.MetadataEntry({
-            key: "owner",
-            value: abi.encode(deployer)
-        });
-        
+        metadata[0] = IIdentityRegistry.MetadataEntry({key: "name", value: abi.encode(agentName)});
+        metadata[1] = IIdentityRegistry.MetadataEntry({key: "type", value: abi.encode(agentType)});
+        metadata[2] = IIdentityRegistry.MetadataEntry({key: "owner", value: abi.encode(deployer)});
+
         uint256 agentId = registry.register(agentUri, metadata);
-        
+
         vm.stopBroadcast();
-        
+
         console.log("==========================================");
         console.log("AGENT REGISTERED!");
         console.log("==========================================");
@@ -101,24 +91,24 @@ contract QueryAgent is Script {
     function run() external view {
         address payable registryAddress = payable(vm.envAddress("IDENTITY_REGISTRY"));
         uint256 agentId = vm.envUint("AGENT_ID");
-        
+
         IdentityRegistry registry = IdentityRegistry(registryAddress);
-        
+
         console.log("==========================================");
         console.log("Agent Information");
         console.log("==========================================");
         console.log("Registry:", registryAddress);
         console.log("Agent ID:", agentId);
         console.log("");
-        
+
         // Basic info
         address owner = registry.ownerOf(agentId);
         string memory uri = registry.tokenURI(agentId);
-        
+
         console.log("Owner:", owner);
         console.log("Token URI:", uri);
         console.log("");
-        
+
         // Try to get common metadata
         console.log("Metadata:");
         string[] memory keys = new string[](5);
@@ -127,8 +117,8 @@ contract QueryAgent is Script {
         keys[2] = "type";
         keys[3] = "model";
         keys[4] = "version";
-        
-        for (uint i = 0; i < keys.length; i++) {
+
+        for (uint256 i = 0; i < keys.length; i++) {
             try registry.getMetadata(agentId, keys[i]) returns (bytes memory value) {
                 if (value.length > 0) {
                     string memory decoded = abi.decode(value, (string));
@@ -136,8 +126,7 @@ contract QueryAgent is Script {
                 }
             } catch {}
         }
-        
+
         console.log("==========================================");
     }
 }
-
