+++ packages/contracts/test/NodeStakingEdgeCases.t.sol
@@ -15,40 +15,34 @@ contract NodeStakingEdgeCasesTest is Test {
     MockTokenRegistry public registry;
     MockPaymasterFactory public factory;
     MockPriceOracle public oracle;
-    
+
     address public alice = address(0xA11CE);
     address public oracleAddr = address(0x04AC1E);
     address public owner = address(this);
-    
+
     function setUp() public {
         token = new MockERC20("TEST", "TEST", 18);
         registry = new MockTokenRegistry();
         factory = new MockPaymasterFactory();
         oracle = new MockPriceOracle();
-        
+
         registry.register(address(token));
         factory.setPaymaster(address(token), address(0xF4ED));
         oracle.setPrice(address(token), 1 ether);
-        
-        staking = new NodeStakingManager(
-            address(registry),
-            address(factory),
-            address(oracle),
-            oracleAddr,
-            owner
-        );
-        
+
+        staking = new NodeStakingManager(address(registry), address(factory), address(oracle), oracleAddr, owner);
+
         token.mint(alice, 100000 ether);
         token.mint(address(staking), 1000000 ether);
         vm.deal(address(staking), 100 ether);
     }
-    
+
     function testFirstStakeWithZeroTotalStaked() public {
         // First registration when totalStakedUSD = 0
         vm.startPrank(alice);
-        
+
         token.approve(address(staking), 10000 ether);
-        
+
         // Should not revert on division by zero
         bytes32 nodeId = staking.registerNode(
             address(token),
@@ -57,63 +51,55 @@ contract NodeStakingEdgeCasesTest is Test {
             "https://first-node.com",
             INodeStakingManager.Region.NorthAmerica
         );
-        
+
         (INodeStakingManager.NodeStake memory node,,) = staking.getNodeInfo(nodeId);
         assertTrue(node.isActive);
-        
+
         vm.stopPrank();
     }
-    
+
     function testDiversityBonusWithZeroTotalStaked() public {
         // Enable diversity bonus with no stakes
         vm.prank(owner);
         staking.enableTokenDiversityBonus(true);
-        
+
         vm.startPrank(alice);
         token.approve(address(staking), 10000 ether);
-        
+
         // Should not revert even with diversity bonus enabled
         bytes32 nodeId = staking.registerNode(
-            address(token),
-            10000 ether,
-            address(token),
-            "https://test.com",
-            INodeStakingManager.Region.Asia
+            address(token), 10000 ether, address(token), "https://test.com", INodeStakingManager.Region.Asia
         );
-        
+
         // Should calculate rewards without division by zero
         vm.warp(block.timestamp + 30 days);
         uint256 rewards = staking.calculatePendingRewards(nodeId);
         assertGt(rewards, 0);
-        
+
         vm.stopPrank();
     }
-    
+
     function testOracleReturnsZeroPrice() public {
         vm.startPrank(alice);
-        
+
         token.approve(address(staking), 10000 ether);
-        
+
         // Set price to zero
         oracle.setPrice(address(token), 0);
-        
+
         // Should revert with proper error
         vm.expectRevert();
         staking.registerNode(
-            address(token),
-            10000 ether,
-            address(token),
-            "https://test.com",
-            INodeStakingManager.Region.NorthAmerica
+            address(token), 10000 ether, address(token), "https://test.com", INodeStakingManager.Region.NorthAmerica
         );
-        
+
         vm.stopPrank();
     }
-    
+
     function testZeroStakeAmount() public {
         vm.startPrank(alice);
         token.approve(address(staking), 0);
-        
+
         vm.expectRevert();
         staking.registerNode(
             address(token),
@@ -122,13 +108,13 @@ contract NodeStakingEdgeCasesTest is Test {
             "https://test.com",
             INodeStakingManager.Region.NorthAmerica
         );
-        
+
         vm.stopPrank();
     }
-    
+
     function testZeroAddressToken() public {
         vm.startPrank(alice);
-        
+
         vm.expectRevert();
         staking.registerNode(
             address(0), // Zero address
@@ -137,46 +123,42 @@ contract NodeStakingEdgeCasesTest is Test {
             "https://test.com",
             INodeStakingManager.Region.NorthAmerica
         );
-        
+
         vm.stopPrank();
     }
-    
+
     function testInternalClaimPaysFees() public {
         vm.startPrank(alice);
-        
+
         token.approve(address(staking), 10000 ether);
         bytes32 nodeId = staking.registerNode(
-            address(token),
-            10000 ether,
-            address(token),
-            "https://test.com",
-            INodeStakingManager.Region.NorthAmerica
+            address(token), 10000 ether, address(token), "https://test.com", INodeStakingManager.Region.NorthAmerica
         );
-        
+
         vm.stopPrank();
-        
+
         // Simulate time + performance
         vm.warp(block.timestamp + 30 days);
         vm.prank(oracleAddr);
         staking.updatePerformance(nodeId, 10000, 1000000, 50);
-        
+
         // Deregister (calls internal claim)
         address paymaster = factory.getPaymaster(address(token));
         uint256 paymasterBalBefore = paymaster.balance;
-        
+
         vm.warp(block.timestamp + 7 days);
         vm.prank(alice);
         staking.deregisterNode(nodeId);
-        
+
         uint256 paymasterBalAfter = paymaster.balance;
-        
+
         // Paymaster should have received ETH from internal claim
         assertGt(paymasterBalAfter, paymasterBalBefore, "Paymaster should receive fees from internal claim");
     }
-    
+
     function testGetNetworkStatsNoDOS() public {
         // Register many nodes
-        for (uint i = 0; i < 50; i++) {
+        for (uint256 i = 0; i < 50; i++) {
             token.mint(address(uint160(0x1000 + i)), 100000 ether);
             vm.startPrank(address(uint160(0x1000 + i)));
             token.approve(address(staking), 10000 ether);
@@ -189,7 +171,7 @@ contract NodeStakingEdgeCasesTest is Test {
             );
             vm.stopPrank();
         }
-        
+
         // Should not run out of gas
         (uint256 totalNodes,,) = staking.getNetworkStats();
         assertEq(totalNodes, 50);
@@ -203,32 +185,32 @@ contract MockERC20 {
     uint8 public decimals;
     mapping(address => uint256) private balances;
     mapping(address => mapping(address => uint256)) private allowances;
-    
+
     constructor(string memory _name, string memory _symbol, uint8 _decimals) {
         name = _name;
         symbol = _symbol;
         decimals = _decimals;
     }
-    
+
     function mint(address to, uint256 amount) external {
         balances[to] += amount;
     }
-    
+
     function balanceOf(address account) external view returns (uint256) {
         return balances[account];
     }
-    
+
     function transfer(address to, uint256 amount) external returns (bool) {
         balances[msg.sender] -= amount;
         balances[to] += amount;
         return true;
     }
-    
+
     function approve(address spender, uint256 amount) external returns (bool) {
         allowances[msg.sender][spender] = amount;
         return true;
     }
-    
+
     function transferFrom(address from, address to, uint256 amount) external returns (bool) {
         allowances[from][msg.sender] -= amount;
         balances[from] -= amount;
@@ -240,18 +222,18 @@ contract MockERC20 {
 contract MockTokenRegistry {
     mapping(address => bool) public registered;
     address[] public tokens;
-    
+
     function register(address token) external {
         if (!registered[token]) {
             registered[token] = true;
             tokens.push(token);
         }
     }
-    
+
     function isRegistered(address token) external view returns (bool) {
         return registered[token];
     }
-    
+
     function getAllTokens() external view returns (address[] memory) {
         return tokens;
     }
@@ -259,15 +241,15 @@ contract MockTokenRegistry {
 
 contract MockPaymasterFactory {
     mapping(address => address) public paymasters;
-    
+
     function setPaymaster(address token, address paymaster) external {
         paymasters[token] = paymaster;
     }
-    
+
     function hasPaymaster(address token) external view returns (bool) {
         return paymasters[token] != address(0);
     }
-    
+
     function getPaymaster(address token) external view returns (address) {
         return paymasters[token];
     }
@@ -275,13 +257,12 @@ contract MockPaymasterFactory {
 
 contract MockPriceOracle {
     mapping(address => uint256) public prices;
-    
+
     function setPrice(address token, uint256 price) external {
         prices[token] = price;
     }
-    
+
     function getPrice(address token) external view returns (uint256) {
         return prices[token];
     }
 }
-
